Allow specifying the version name with Renv install
Currently, when using Renv install, the name of the version stored in ~/.Renv/versions is taken from the version number, such as 3.2.3. This adds support for changing the version name. This is useful when creating versions for different projects, where you might want the same version number but want to isolate the packages installed into those versions.

This feature is already available when running R-build as a standalone program. Here, I'm adding the feature to R-build when called as an Renv plugin.